+++ sources/libhxcadaptor.h
@@ -42,9 +42,9 @@ int hxc_fclose(FILE * f);
 #endif
 int hxc_stat( const char *filename, struct stat *buf);
 
-long hxc_find_first_file(char *folder,char *file,filefoundinfo* fileinfo);
-long hxc_find_next_file(long handleff,char *folder,char *file,filefoundinfo* fileinfo);
-long hxc_find_close(long handle);
+int hxc_find_first_file(char *folder,char *file,filefoundinfo* fileinfo);
+int hxc_find_next_file(int handleff,char *folder,char *file,filefoundinfo* fileinfo);
+int hxc_find_close(int handle);
 
 int  hxc_mkdir(char * folder);
